requirements/test.txt
south
generic-request-signer>=0.0.8
tox
mock
django-nose
nosexcover
coverage
lettuce
